CLi create switched to new utils. Some of utils reorganized.
