+++ aea/cli/utils/config.py
@@ -22,12 +22,17 @@
 import logging
 import os
 from pathlib import Path
+from typing import Dict
 
 import click
 
 import jsonschema  # type: ignore
 
+import yaml
+
+from aea.cli.utils.constants import CLI_CONFIG_PATH
 from aea.cli.utils.context import Context
+from aea.cli.utils.package_utils import _load_yaml
 from aea.configurations.base import DEFAULT_AEA_CONFIG_FILE
 
 
@@ -65,3 +70,43 @@ def try_to_load_agent_config(
                     DEFAULT_AEA_CONFIG_FILE
                 )
             )
+
+
+def _init_cli_config() -> None:
+    """
+    Create cli config folder and file.
+
+    :return: None
+    """
+    conf_dir = os.path.dirname(CLI_CONFIG_PATH)
+    if not os.path.exists(conf_dir):
+        os.makedirs(conf_dir)
+    with open(CLI_CONFIG_PATH, "w+") as f:
+        yaml.dump({}, f, default_flow_style=False)
+
+
+def _update_cli_config(dict_conf: Dict) -> None:
+    """
+    Update CLI config and write to yaml file.
+
+    :param dict_conf: dict config to write.
+
+    :return: None
+    """
+    config = _get_or_create_cli_config()
+    config.update(dict_conf)
+    with open(CLI_CONFIG_PATH, "w") as f:
+        yaml.dump(config, f, default_flow_style=False)
+
+
+def _get_or_create_cli_config() -> Dict:
+    """
+    Read or create CLI config from yaml file.
+
+    :return: dict CLI config.
+    """
+    try:
+        return _load_yaml(CLI_CONFIG_PATH)
+    except FileNotFoundError:
+        _init_cli_config()
+    return _load_yaml(CLI_CONFIG_PATH)
